Moved historic dictionary code to sub-module
